# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: Auth
    description: Browser authentication endpoints.
  - name: Calendar
    description: Google Calendar sync endpoints.
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /auth/user:
    get:
      tags: [Auth]
      operationId: getCurrentAuthUser
      summary: Get the currently authenticated user
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '200':
          description: Auth status resolved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthUserEnvelope'
  /auth/register:
    post:
      tags: [Auth]
      operationId: registerWithPassword
      summary: Register with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Account created and session started.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthUserEnvelope'
        '400':
          description: Validation error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: Email already registered.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /auth/login:
    post:
      tags: [Auth]
      operationId: loginWithPassword
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authenticated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthUserEnvelope'
        '401':
          description: Invalid credentials.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '429':
          description: Too many attempts or account locked.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /auth/google:
    get:
      tags: [Auth]
      operationId: beginGoogleLogin
      summary: Start Google OAuth login flow
      parameters:
        - name: returnTo
          in: query
          required: false
          schema:
            type: string
      responses:
        '302':
          description: Redirect to Google authorization page.
  /auth/google/callback:
    get:
      tags: [Auth]
      operationId: handleGoogleCallback
      summary: Complete Google OAuth login flow
      responses:
        '302':
          description: Redirect to app on success or error page on failure.
  /auth/logout:
    post:
      tags: [Auth]
      operationId: logoutSession
      summary: Clear the session cookie
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '200':
          description: Logged out.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogoutSuccess'
  /calendar/sync:
    get:
      tags: [Calendar]
      operationId: syncGoogleCalendar
      summary: Fetch next 30 days of events from the user's primary Google Calendar
      responses:
        '200':
          description: Events fetched and ready to merge.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalendarSyncResponse'
        '401':
          description: Not authenticated or token expired.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalendarErrorEnvelope'
        '403':
          description: No Google token — user must sign in with Google.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalendarErrorEnvelope'
  /calendar/push:
    post:
      tags: [Calendar]
      operationId: pushEventToGoogleCalendar
      summary: Create a new event in the user's primary Google Calendar
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CalendarPushRequest'
      responses:
        '200':
          description: Event created in Google Calendar.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalendarPushResponse'
        '400':
          description: Missing required fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: No Google token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalendarErrorEnvelope'
components:
  parameters:
    AuthorizationSessionHeader:
      name: Authorization
      in: header
      required: false
      description: Opaque session token — `Bearer <sid>`.
      schema:
        type: string
    SessionCookie:
      name: sid
      in: cookie
      required: false
      description: Browser session cookie.
      schema:
        type: string
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    AuthUser:
      type: object
      required: [id, email, firstName, lastName, profileImageUrl]
      properties:
        id:
          type: string
        email:
          type: [string, 'null']
          format: email
        firstName:
          type: [string, 'null']
        lastName:
          type: [string, 'null']
        profileImageUrl:
          type: [string, 'null']
    AuthUserEnvelope:
      type: object
      required: [user]
      properties:
        user:
          oneOf:
            - $ref: '#/components/schemas/AuthUser'
            - type: 'null'
    RegisterRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        firstName:
          type: string
        lastName:
          type: string
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    LogoutSuccess:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          const: true
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
    CalendarEvent:
      type: object
      required: [id, title, date, time, color, source, googleId, allDay]
      properties:
        id:
          type: string
        title:
          type: string
        date:
          type: string
          description: YYYY-MM-DD
        time:
          type: string
          description: HH:MM (empty string for all-day events)
        color:
          type: string
        source:
          type: string
          enum: [google]
        googleId:
          type: string
        allDay:
          type: boolean
    CalendarSyncResponse:
      type: object
      required: [events, syncedAt]
      properties:
        events:
          type: array
          items:
            $ref: '#/components/schemas/CalendarEvent'
        syncedAt:
          type: string
          format: date-time
    CalendarPushRequest:
      type: object
      required: [title, date]
      properties:
        title:
          type: string
        date:
          type: string
          description: YYYY-MM-DD
        time:
          type: string
          description: HH:MM (omit for all-day)
    CalendarPushResponse:
      type: object
      required: [googleId]
      properties:
        googleId:
          type: [string, 'null']
        htmlLink:
          type: [string, 'null']
    CalendarErrorEnvelope:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          description: Machine-readable code (NO_GOOGLE_TOKEN, TOKEN_EXPIRED, CALENDAR_API_ERROR)
        message:
          type: string